split other_views.py into multiple files
