+++ daas/daas_app/api/__init__.py
@@ -3,4 +3,5 @@ from .get_token import get_token_view
 from .reprocess import ReprocessAPIView
 from .upload import UploadAPIView
 from .download_sample import download_sample_view
+from .download_source_code import download_source_code_view
 from . import internals
