+++ daas/daas_app/views/__init__.py
@@ -1,3 +1,7 @@
+from .cancel_job import cancel_job_view
 from .charts import (SamplesPerType, SamplesPerSize, SamplesPerElapsedTime, SamplesPerUploadDate, SamplesPerProcessDate,
                      SamplesPerStatusForFileType)
-from .other_views import *
+from .index import IndexView
+from .reprocess import reprocess_view
+from .sample_delete import SampleDeleteView
+from .upload_file import upload_file_view, file_already_uploaded_view, no_filter_found_view
